requirements.txt
fastapi>=0.110.0
mangum>=0.17.0
python-multipart>=0.0.9
pillow>=10.3.0
